+++ tests/files/no-named-as-default/exports.js
@@ -1,0 +1,4 @@
+const variable = 1;
+
+export { variable };
+export default variable;