+++ tests/files/no-named-as-default/re-exports.js
@@ -1,0 +1,2 @@
+export { something as default } from "./something";
+export { something } from "./something";